+++ core/pkg/artifacts/saver.go
@@ -177,7 +177,7 @@ func (as *ArtifactSaver) createArtifact() (
 	}
 
 	// Check which fields are actually supported on the input
-	inputFieldNames, err := getInputFields(as.ctx, as.graphqlClient, "CreateArtifactInput")
+	inputFieldNames, err := GetGraphQLInputFields(as.ctx, as.graphqlClient, "CreateArtifactInput")
 	if err != nil {
 		return gql.CreatedArtifactArtifact{}, err
 	}
@@ -216,23 +216,6 @@ func (as *ArtifactSaver) createArtifact() (
 		return gql.CreatedArtifactArtifact{}, err
 	}
 	return response.GetCreateArtifact().GetArtifact(), nil
-}
-
-func getInputFields(ctx context.Context, client graphql.Client, typeName string) ([]string, error) {
-	response, err := gql.InputFields(ctx, client, typeName)
-	if err != nil {
-		return nil, err
-	}
-	typeInfo := response.GetTypeInfo()
-	if typeInfo == nil {
-		return nil, fmt.Errorf("unable to verify allowed fields for %s", typeName)
-	}
-	fields := typeInfo.GetInputFields()
-	fieldNames := make([]string, len(fields))
-	for i, field := range fields {
-		fieldNames[i] = field.GetName()
-	}
-	return fieldNames, nil
 }
 
 type createArtifactManifest = gql.CreateArtifactManifestCreateArtifactManifestCreateArtifactManifestPayloadArtifactManifest